pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asset-cluster-migration"
version = "0.4.0"
description = "Dynamic multi-layer asset topology and cluster migration under geopolitical stress"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "networkx>=3.2",
    "python-igraph>=0.11",
    "leidenalg>=0.10",
    "hmmlearn>=0.3",
    "ruptures>=1.1",
    "dcor>=0.7",
    "POT>=0.9",
    "pyarrow>=14.0",
    "httpx>=0.27",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "rich>=13.7",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "nbstripout>=0.7",
    "detect-secrets>=1.4",
    "ipykernel>=6.29",
]

[project.scripts]
acm = "src.pipeline.orchestrator:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
